robot/src/requirements.txt
robotframework
docker-py
paramiko
